requirements.txt
pycryptodome==3.9.7
requests==2.23.0
